+++ src/schick/seg099.cpp
@@ -224,7 +224,7 @@ void spell_illusionen(void)
 		}
 	} else {
 		/* print a failure message */
-		sprintf(g_dtp2,	get_tx(3), GUI_names_grammar((signed int)0x8000, g_spelltarget_e->mon_id, 1));
+		sprintf(g_dtp2,	get_tx(3), GUI_names_grammar(0x8000, g_spelltarget_e->mon_id, 1));
 
 		/* costs 2 AE */
 		g_spell_special_aecost = 2;
@@ -302,7 +302,7 @@ void spell_band(void)
 		/* set 'tied' flag */
 		g_spelltarget_e->flags.tied = 1;
 
-		sprintf(g_dtp2, get_tx(6), (uint8_t*)GUI_names_grammar((signed int)0x8000, g_spelltarget_e->mon_id, 1));
+		sprintf(g_dtp2, get_tx(6), (uint8_t*)GUI_names_grammar(0x8000, g_spelltarget_e->mon_id, 1));
 	} else {
 		/* cast hero */
 
@@ -350,7 +350,7 @@ void spell_bannbaladin(void)
 		/* set 'tame' flag */
 		g_spelltarget_e->flags.tame = 1;
 
-		sprintf(g_dtp2,	get_tx(9), (uint8_t*)GUI_names_grammar((signed int)0x8000, g_spelltarget_e->mon_id, 1));
+		sprintf(g_dtp2,	get_tx(9), (uint8_t*)GUI_names_grammar(0x8000, g_spelltarget_e->mon_id, 1));
 	}
 }
 
@@ -373,7 +373,7 @@ void spell_boeser_blick(void)
 		g_spelltarget_e->attacks = 2;
 
 		/* prepare message */
-		sprintf(g_dtp2,	get_tx(10), (char*)GUI_names_grammar((signed int)0x8000, g_spelltarget_e->mon_id, 1));
+		sprintf(g_dtp2,	get_tx(10), (char*)GUI_names_grammar(0x8000, g_spelltarget_e->mon_id, 1));
 	}
 }
 
@@ -429,7 +429,7 @@ void spell_herrdertiere(void)
 			/* set 'tame' flag */
 			g_spelltarget_e->flags.tame = 1;
 
-			sprintf(g_dtp2,	get_tx(9), (uint8_t*)GUI_names_grammar((signed int)0x8000, g_spelltarget_e->mon_id, 1));
+			sprintf(g_dtp2,	get_tx(9), (uint8_t*)GUI_names_grammar(0x8000, g_spelltarget_e->mon_id, 1));
 		}
 	}
 }
@@ -450,7 +450,7 @@ void spell_horriphobus(void)
 		/* unset 'renegade' flag */
 		g_spelltarget_e->flags.renegade = 0;
 
-		sprintf(g_dtp2,	get_tx(12), (uint8_t*)GUI_names_grammar((signed int)0x8000, g_spelltarget_e->mon_id, 1));
+		sprintf(g_dtp2,	get_tx(12), (uint8_t*)GUI_names_grammar(0x8000, g_spelltarget_e->mon_id, 1));
 	}
 }
 
@@ -492,7 +492,7 @@ void spell_somnigravis(void)
 		g_spelltarget_e->flags.asleep = 1;
 
 		/* prepare message */
-		sprintf(g_dtp2,	get_tx(13), (uint8_t*)GUI_names_grammar((signed int)0x8000, g_spelltarget_e->mon_id, 1));
+		sprintf(g_dtp2,	get_tx(13), (uint8_t*)GUI_names_grammar(0x8000, g_spelltarget_e->mon_id, 1));
 		return;
 	}
 
@@ -533,7 +533,7 @@ void spell_zwingtanz(void)
 		g_spelltarget_e->flags.dancing = 1;
 
 		/* prepare message */
-		sprintf(g_dtp2,	get_tx(14), (uint8_t*)GUI_names_grammar((signed int)0x8000, g_spelltarget_e->mon_id, 1));
+		sprintf(g_dtp2,	get_tx(14), (uint8_t*)GUI_names_grammar(0x8000, g_spelltarget_e->mon_id, 1));
 	}
 }
 
@@ -612,14 +612,14 @@ void spell_skelettarius(void)
 	if (!g_spelltarget_e->flags.dead) {
 
 		/* prepare message */
-		sprintf(g_dtp2,	get_tx(15), (uint8_t*)GUI_names_grammar((signed int)0x8000, g_spelltarget_e->mon_id, 1));
+		sprintf(g_dtp2,	get_tx(15), (uint8_t*)GUI_names_grammar(0x8000, g_spelltarget_e->mon_id, 1));
 
 		/* set ae costs */
 		g_spell_special_aecost = 0;
 	} else {
 
 		/* prepare message */
-		sprintf(g_dtp2,	get_tx(16), (uint8_t*)GUI_names_grammar((signed int)0x8000, g_spelltarget_e->mon_id, 1));
+		sprintf(g_dtp2,	get_tx(16), (uint8_t*)GUI_names_grammar(0x8000, g_spelltarget_e->mon_id, 1));
 
 		fighter = FIG_get_fighter(g_spelltarget_e->fighter_id);
